Update README.md
Recently, I was working with a billing system in python. In the last stages of the application, I wanted to add a barcode reading function to my code. Then I searched it up in the internet but, couldn't find a better solution. As an undergraduate software engineer, I had a very lack knowledge about barcode system and barcode readers. Then I had to do my own research on barcodes and with that knowledge made this simple but complete barcode function. Hope this would help for knowledge explorers and this might save your time :)